Update Map.java
(1) I think it is real critical to have the HashMap<Point, ArrayList<Organism>> as a field variable. Each object should keep it for the duration of its existence. (see lines 16; 24; 35; 36)

(2) Creating an Array of Organisms is not needed here. Map does not require an Organism collection of its own.. (line 37)